Update 01-declarative-copilot.md
corrects what needs to be done with the provided image. The instruction said one needs to copy the image and replace it (the copied image) with the image in the project. IN fact though, it is the image in the project that needs to be replaces by the image provided on the website.